Replace ES6 Object.assign method with polyfill
Object.assign is an ES6 method and is not supported on older browsers such
as Internet Explorer and the Android mobile browser. This commit adds
a commonly-used npm module for a polyfill to fall back on if the user
is using an older browser. If the user has an evergreen browser the
native Object.assign method is used.

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Object/assign